Add wait period for the tunnel connection health check
