+++ pkg/operation/botanist/tunnel.go
@@ -17,8 +17,12 @@ package botanist
 import (
 	"context"
 	"fmt"
+	"time"
 
 	"github.com/gardener/gardener/pkg/client/kubernetes"
+	"github.com/gardener/gardener/pkg/operation/common"
+	"github.com/gardener/gardener/pkg/operation/shoot"
+	kutil "github.com/gardener/gardener/pkg/utils/kubernetes"
 	"github.com/gardener/gardener/pkg/utils/retry"
 
 	"github.com/sirupsen/logrus"
@@ -54,4 +58,48 @@ func CheckTunnelConnection(ctx context.Context, shootClient kubernetes.Interface
 
 	logger.Info("Tunnel connection has been established.")
 	return retry.Ok()
+}
+
+// CheckAndWaitForTunnelConnection checks until the tunnel connection between the control plane and the shoot networks
+// is established, or until the configured timeout.
+func CheckAndWaitForTunnelConnection(ctx context.Context, shootClient kubernetes.Interface, shoot *shoot.Shoot, logger logrus.FieldLogger, tunnelName string, timeout time.Duration) error {
+	timeoutCtx, cancel := context.WithTimeout(ctx, timeout)
+	defer cancel()
+
+	return retry.Until(timeoutCtx, 5*time.Second, func(ctx context.Context) (bool, error) {
+		done, err := CheckTunnelConnection(ctx, shootClient, logger, common.VPNTunnel)
+
+		// If the tunnel connection check failed but is not yet "done" (i.e., will be retried, hence, it didn't fail
+		// with a severe error), and if the classic VPN solution is used for the shoot cluster then let's try to fetch
+		// the last events of the vpn-shoot service (potentially indicating an error with the load balancer service).
+		if err != nil &&
+			!done &&
+			!shoot.ReversedVPNEnabled {
+
+			logger.Errorf("error %v occurred while checking the tunnel connection", err)
+
+			service := &corev1.Service{
+				TypeMeta: metav1.TypeMeta{
+					APIVersion: corev1.SchemeGroupVersion.String(),
+					Kind:       "Service",
+				},
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "vpn-shoot",
+					Namespace: metav1.NamespaceSystem,
+				},
+			}
+
+			eventsErrorMessage, err2 := kutil.FetchEventMessages(ctx, shootClient.Client().Scheme(), shootClient.Client(), service, corev1.EventTypeWarning, 2)
+			if err2 != nil {
+				logger.Errorf("error %v occurred while fetching events for VPN load balancer service", err2)
+				return retry.SevereError(fmt.Errorf("'%w' occurred but could not fetch events for more information", err))
+			}
+
+			if eventsErrorMessage != "" {
+				return retry.SevereError(fmt.Errorf("%s\n\n%s", err.Error(), eventsErrorMessage))
+			}
+		}
+
+		return done, err
+	})
 }
